python/requirements.txt
kagglehub
pandas
langchain
tqdm
openai
supabase
python-dotenv
numpy